refactor: reorganize codebase with multi-stage Docker builds and improved server initialization
