+++ .docker/redis/redis.conf
@@ -1,0 +1,14 @@
+# Redis configuration for Docker environment
+# Reduce log verbosity
+loglevel warning
+
+# Configure saves with less frequent snapshots
+save 900 1
+save 300 10
+save 60 10000
+
+# Disable some features that require memory overcommit
+stop-writes-on-bgsave-error no
+
+# Memory management
+maxmemory-policy allkeys-lru
