+++ .env.prod.example
@@ -1,0 +1,35 @@
+# Production Environment Variables
+# Copy this to .env.prod and fill in your production values
+
+NODE_ENV=production
+
+# Backend URLs
+REMOTE_BACKEND_HOST_URL=https://api.yourdomain.com
+LOCAL_BACKEND_HOST_URL=http://api:4005
+FRONTEND_REDIRECT_URL=https://yourdomain.com
+
+# Discord OAuth
+DISCORD_OAUTH_REDIRECT_URL=https://api.yourdomain.com/auth/discord/callback
+DISCORD_CLIENT_ID=your_discord_client_id
+DISCORD_AUTH_CLIENT_ID=your_discord_auth_client_id
+DISCORD_AUTH_CLIENT_SECRET=your_discord_auth_client_secret
+DISCORD_OAUTH_SCOPES=identify email guilds
+
+# JWT Configuration
+JWT_SECRET=your_strong_random_jwt_secret_here
+JWT_MAX_AGE=86400000
+
+# CORS Configuration
+CORS_ORIGIN_LIST=https://yourdomain.com,https://www.yourdomain.com
+
+# Database (Production)
+PG_DB_URL_LOCAL=postgresql://postgres:your_strong_password@db:5432/taskcord
+PG_DB_URL_PROD=postgresql://postgres:your_strong_password@db:5432/taskcord
+POSTGRES_PASSWORD=your_strong_password
+
+# Redis (Production)
+REDIS_URL_LOCAL=redis://redis:6379
+REDIS_URL_PROD=redis://redis:6379
+
+# Application Port
+PORT=4005
