+++ Dockerfile.dbot
@@ -1,0 +1,23 @@
+# apps/discord-bot/Dockerfile
+FROM node:22-alpine
+
+WORKDIR /app
+
+RUN npm install -g pnpm
+
+# --- Step 1: Install Dependencies (Cached) ---
+COPY package.json pnpm-lock.yaml pnpm-workspace.yaml ./
+COPY apps/api/package.json ./apps/api/
+COPY apps/dashboard-client/package.json ./apps/dashboard-client/
+COPY apps/discord-bot/package.json ./apps/discord-bot/
+COPY packages/*/package.json ./packages/*/
+
+RUN pnpm install --no-frozen-lockfile
+
+# --- Step 2: Copy Source Code ---
+# The bot might need the database package to interact with the DB
+COPY apps/discord-bot/ ./apps/discord-bot/
+COPY packages/database/ ./packages/database/
+
+# The command to start the bot in development mode
+CMD ["pnpm", "--filter", "discord-bot", "dev"]
