--- packages/redis/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@taskcord/redis",
-  "version": "0.0.0",
-  "private": true,
-  "files": [
-    "dist"
-  ],
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "dev": "tsc --watch",
-    "build": "tsup --dts"
-  },
-  "devDependencies": {
-    "typescript": "latest"
-  },
-  "dependencies": {
-    "ioredis": "^5.5.0"
-  }
-}
