--- packages/redis/tsconfig.json
@@ -1,22 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "display": "Node 20",
-  "_version": "20.1.0",
-
-  "compilerOptions": {
-    "lib": ["es2023", "dom"],
-    "module": "node16",
-    "target": "es2022",
-    "outDir": "dist",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "moduleResolution": "node16",
-    "rootDir": "src",
-    "declaration": true,
-    "noEmit": false,
-    "forceConsistentCasingInFileNames": true
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"]
-}
